+++ _boards/mcxn5xxevk/ota_examples/mcuboot_opensource/cm33_core0/reconfig.cmake
@@ -1,18 +1,12 @@
-
-
 mcux_add_source(
     BASE_PATH ${SdkRootDirPath}
     SOURCES ${board_root}/${board}/ota_examples/mcuboot_opensource/mbedtls_user_config.h
 )
 
-
 mcux_add_include(
     BASE_PATH ${SdkRootDirPath}
     INCLUDES ${board_root}/${board}/ota_examples/mcuboot_opensource
 )
-
-
-
 
 # Add additional configuration
 mcux_add_macro(
@@ -48,18 +42,5 @@ mcux_remove_mdk_linker_script(
     LINKER ${device_root}/MCX/MCXN/MCXN547/arm/MCXN547_cm33_core0_flash.scf
 )
 
-mcux_add_iar_linker_script(
-    BASE_PATH ${SdkRootDirPath}
-    TARGETS debug release
-    LINKER ${board_root}/${board}/ota_examples/mcuboot_opensource/linker/mcxn10_cm33_flash.icf
-)
-mcux_add_armgcc_linker_script(
-    BASE_PATH ${SdkRootDirPath}
-    TARGETS debug release
-    LINKER ${board_root}/${board}/ota_examples/mcuboot_opensource/linker/mcxn10_cm33_flash.ld
-)
-mcux_add_mdk_linker_script(
-    BASE_PATH ${SdkRootDirPath}
-    TARGETS debug release
-    LINKER ${board_root}/${board}/ota_examples/mcuboot_opensource/linker/mcxn10_cm33_flash.scf
-)
+# Specific linkers are added using project segments
+# see examples/_common/project_segments/ota_examples/mcxn_boards/prjseg.cmake