+++ ota_examples/CHANGELOG.md
@@ -6,9 +6,10 @@ All notable changes to MCUboot fork in MCUXpresso SDK and related ota examples w
 
 ### Added
 - Added `ota_rom_basic` example demonstrating ROM bootloader processing a secure binary as an OTA image (see documentation in docs folder)
-- Enabled `ota_rom_basic` example for FRDM-MCXN947, MCXN9XXEVK, FRDM-RW612 and RD-RW612BGA
+- Enabled `ota_rom_basic` example for FRDM-MCXN947, MCX-N9XX-EVK, FRDM-RW612 and RD-RW612BGA
 - Enabled flash remap feature for MIMXRT700EVK
 - FRDM-MCXE247 basic MCUboot support
+- FRDM-MCXN947, MCX-N5XX-EVK and MCX-N9XX-EVK supports alternative build using MCUboot located in the internal flash instead of the IFR region (default) - see board readme for more information
 
 ### Removed
 - Removed `ota_rom_server` (replaced by simpler `ota_rom_basic`)